Gamy/requirements.txt
python-dotenv==1.2.1
asyncio==4.0.0
discord.py==2.6.4
